Reorganise commands and split up larger files
