+++ src/backend/common/data/commands/__init__.py
@@ -1,21 +1,8 @@
-# re-export all the commands
 from .common import *
 from .auth import *
-from .command_log import *
-from .db_admin import *
-from .friend_codes import *
+from .system import *
 from .moderation import *
-from .notifications import *
 from .players import *
-from .roles import *
-from .s3 import *
-from .squads import *
 from .teams import *
-from .tournament_placements import *
-from .tournament_registrations import *
-from .tournament_series import *
-from .tournament_templates import *
 from .tournaments import *
-from .user_settings import *
 from .users import *
-from .player_bans import *
