+++ src/backend/common/data/commands/auth/__init__.py
@@ -1,0 +1,7 @@
+from .discord import *
+from .permissions import *
+from .roles import *
+from .seriesroles import *
+from .sessions import *
+from .teamroles import *
+from .tournamentroles import *
