+++ src/backend/common/data/commands/auth/sessions.py
@@ -1,0 +1,61 @@
+from dataclasses import dataclass
+from common.data.commands import Command
+from common.data.models import *
+
+@dataclass 
+class GetUserIdFromSessionCommand(Command[User | None]):
+    session_id: str
+
+    async def handle(self, db_wrapper, s3_wrapper):
+        async with db_wrapper.connect() as db:
+            async with db.execute("SELECT user_id FROM sessions WHERE session_id = ?", (self.session_id,)) as cursor:
+                row = await cursor.fetchone()
+
+            if row is None:
+                return None
+
+            user_id = int(row[0])
+            async with db.execute("SELECT player_id FROM users WHERE id = ?", (user_id,)) as cursor:
+                row = await cursor.fetchone()
+            assert row is not None
+            player_id = row[0]
+            
+            return User(user_id, player_id)
+            
+@dataclass
+class IsValidSessionCommand(Command[bool]):
+    session_id: str
+
+    async def handle(self, db_wrapper, s3_wrapper):
+        async with db_wrapper.connect(readonly=True) as db:
+            async with db.execute("SELECT EXISTS(SELECT 1 FROM sessions WHERE session_id = ?)", (self.session_id,)) as cursor:
+                row = await cursor.fetchone()
+
+            return row is not None and bool(row[0])
+        
+@dataclass
+class CreateSessionCommand(Command[None]):
+    session_id: str
+    user_id: int
+    expires_on: int
+
+    async def handle(self, db_wrapper, s3_wrapper):
+        async with db_wrapper.connect() as db:
+            command = "INSERT INTO sessions(session_id, user_id, expires_on) VALUES (?, ?, ?)"
+            async with db.execute(command, (self.session_id, self.user_id, self.expires_on)) as cursor:
+                rows_inserted = cursor.rowcount
+
+            # TODO: Run queries to identify why session creation failed
+            if rows_inserted != 1:
+                raise Problem("Failed to create session")
+                
+            await db.commit()
+            
+@dataclass
+class DeleteSessionCommand(Command[None]):
+    session_id: str
+
+    async def handle(self, db_wrapper, s3_wrapper):
+        async with db_wrapper.connect() as db:
+            await db.execute("DELETE FROM sessions WHERE session_id = ?", (self.session_id, ))
+            await db.commit()
