+++ src/backend/common/data/commands/moderation/__init__.py
@@ -1,0 +1,3 @@
+from .mod_notifications import *
+from .player_bans import *
+from .word_filter import *
