+++ src/backend/common/data/commands/players/__init__.py
@@ -1,0 +1,5 @@
+from .claims import *
+from .friend_codes import *
+from .name_changes import *
+from .notes import *
+from .players import *
