+++ src/backend/common/data/commands/tournaments/__init__.py
@@ -1,0 +1,6 @@
+from .placements import *
+from .registrations import *
+from .series import *
+from .squads import *
+from .templates import *
+from .tournaments import *
