+++ src/backend/common/data/commands/tournaments/registrations.py
