+++ src/backend/common/data/commands/tournaments/squads.py
