+++ src/backend/common/data/commands/tournaments/templates.py
