+++ src/backend/common/data/models/__init__.py
@@ -3,6 +3,7 @@ from .common import *
 from .command_log import *
 from .discord_integration import *
 from .friend_codes import *
+from .mkcv1 import *
 from .moderation import *
 from .notifications import *
 from .players import *
